app.py
"""
Kroger Shopping AI - Streamlit Application
Main entry point for the Python/Streamlit version
"""
import streamlit as st
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Page configuration
st.set_page_config(
    page_title="Kroger Shopping AI",
    page_icon="🛒",
    layout="wide",
    initial_sidebar_state="expanded"
)

# Initialize session state
if 'cart' not in st.session_state:
    st.session_state.cart = []
if 'selected_store' not in st.session_state:
    st.session_state.selected_store = None

# Header
st.markdown("""
<style>
.main-header {
    background: linear-gradient(135deg, #0066cc 0%, #004499 100%);
    padding: 2rem;
    border-radius: 10px;
    margin-bottom: 2rem;
    text-align: center;
}
.main-header h1 {
    color: white;
    margin: 0;
}
</style>
<div class="main-header">
    <h1>🛒 Kroger Shopping AI Assistant</h1>
</div>
""", unsafe_allow_html=True)

# Main content
col1, col2 = st.columns([3, 1])

with col1:
    st.info("🚧 **Migration in Progress** - This is the new Python/Streamlit version of the Kroger Shopping AI")
    
    # Tabs
    tab1, tab2, tab3, tab4 = st.tabs(["🔍 Search Products", "📝 Shopping List", "🛒 Cart", "⚙️ Settings"])
    
    with tab1:
        st.header("Product Search")
        search_term = st.text_input("Search for products", placeholder="Enter product name...")
        if st.button("Search", type="primary"):
            st.success(f"Would search for: {search_term}")
            st.write("Product search functionality will be implemented soon!")
    
    with tab2:
        st.header("Shopping List")
        st.write("Shopping list management coming soon!")
    
    with tab3:
        st.header("Shopping Cart")
        if st.session_state.cart:
            for item in st.session_state.cart:
                st.write(item)
        else:
            st.write("Your cart is empty")
    
    with tab4:
        st.header("Settings")
        store = st.selectbox("Select Store", ["None", "Store 1", "Store 2"])
        if store != "None":
            st.session_state.selected_store = store
            st.success(f"Selected store: {store}")

with col2:
    st.markdown("### Quick Stats")
    st.metric("Cart Items", len(st.session_state.cart))
    st.metric("Selected Store", st.session_state.selected_store or "None")
    
    if st.button("View Progress Tracker"):
        st.markdown("[Open Progress Dashboard](progress.html)")

# Footer
st.markdown("---")
st.markdown("**Status**: Python migration underway. Check progress.html for real-time updates.")