requirements.txt
# Core Dependencies
streamlit==1.28.1
pandas==2.1.1
numpy==1.26.0

# API Client
httpx==0.25.0
tenacity==8.2.3

# Data Validation
pydantic==2.4.2
python-dotenv==1.0.0

# Caching
diskcache==5.6.3
redis==5.0.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Utilities
python-multipart==0.0.6
python-dateutil==2.8.2

# Data Visualization (optional)
plotly==5.17.0
altair==5.1.2

# Development Dependencies
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
mypy==1.5.1
pre-commit==3.4.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.6